Former-commit-id: 87b8c30ce14568c7f38bf711ef9fdd34c0b0003c

+++ Sources/Internal/Render/Shader.cpp
@@ -46,13 +46,14 @@
 
 namespace DAVA
 {
+
 #if defined(__DAVAENGINE_OPENGL__)
 GLuint Shader::activeProgram = 0;
 
 Shader::Shader()
 : RenderResource()
 {
-    DVASSERT(RenderManager::Instance()->GetRenderer() == Core::RENDERER_OPENGL_ES_2_0 || RenderManager::Instance()->GetRenderer() == Core::RENDERER_OPENGL);
+    DVASSERT(RenderManager::Instance()->GetRenderer() == Core::RENDERER_OPENGL_ES_2_0 || RenderManager::Instance()->GetRenderer() == Core::RENDERER_OPENGL || RenderManager::Instance()->GetRenderer() == Core::RENDERER_OPENGL_ES_3_0);
     
     vertexShader = 0;
     fragmentShader = 0;
@@ -61,6 +62,7 @@ Shader::Shader()
     attributeNames = 0;
     activeAttributes = 0;
     activeUniforms = 0;
+    requiredVertexFormat = 0;
     
     //uniforms = 0;
     uniformData = NULL;
@@ -117,7 +119,10 @@ FastName attributeStrings[VERTEX_FORMAT_STREAM_MAX_COUNT] =
     FastName("inTangent"),
     FastName("inBinormal"),
     FastName("inJointWeight"),
-    FastName("inTime")
+    FastName("inTime"),
+    FastName("inPivot"),
+    FastName("inFlexibility"),
+    FastName("inAngleSinCos")
 };
 
 eShaderSemantic Shader::GetShaderSemanticByName(const FastName & name)
@@ -236,6 +241,8 @@ const char * Shader::GetUniformTypeSLName(eUniformType type)
 
 void Shader::ClearLastBindedCaches()
 {
+    //Logger::FrameworkDebug("Frame reset");
+
     for (uint32 k = 0; k < autobindUniformCount; ++k)
     {
         Uniform * currentUniform = autobindUniforms[k];
@@ -273,7 +280,7 @@ Shader::~Shader()
     ReleaseShaderData();
 }
 
-void Shader::ReleaseShaderData()
+void Shader::ReleaseShaderData(bool deleteShader/* = true*/)
 {
     SafeDeleteArray(attributeNames);
     //SafeDeleteArray(uniforms);
@@ -283,7 +290,8 @@ void Shader::ReleaseShaderData()
     SafeRelease(vertexShaderData);
     SafeRelease(fragmentShaderData);
     
-    DeleteShaders();
+    if (deleteShader)
+        DeleteShaders();
 
     activeAttributes = 0;
     activeUniforms = 0;
@@ -318,13 +326,13 @@ void Shader::RecompileInternal(BaseObject * caller, void * param, void *callerDa
     
     if (!CompileShader(&vertexShader, GL_VERTEX_SHADER, vertexShaderDataSize, (GLchar*)vertexShaderDataStart, shaderDefines))
     {
-        Logger::Error("Failed to compile vertex shader: %s", assetName.c_str());
+        Logger::Error("Failed to compile vertex shader: %s defines: %s", assetName.c_str(), shaderDefines.c_str());
         return;
     }
     
     if (!CompileShader(&fragmentShader, GL_FRAGMENT_SHADER, fragmentShaderDataSize, (GLchar*)fragmentShaderDataStart, shaderDefines))
     {
-        Logger::Error("Failed to compile fragment shader: %s", assetName.c_str());
+        Logger::Error("Failed to compile fragment shader: %s defines:%s", assetName.c_str(), shaderDefines.c_str());
         return ;
     }
     
@@ -345,6 +353,9 @@ void Shader::RecompileInternal(BaseObject * caller, void * param, void *callerDa
     char attributeName[512];
     DVASSERT(attributeNames == NULL);
     attributeNames = new FastName[activeAttributes];
+
+    requiredVertexFormat = 0;
+
     for (int32 k = 0; k < activeAttributes; ++k)
     {
         GLint size;
@@ -353,7 +364,12 @@ void Shader::RecompileInternal(BaseObject * caller, void * param, void *callerDa
         attributeNames[k] = FastName(attributeName);
         
         int32 flagIndex = GetAttributeIndexByName(attributeNames[k]);
-        vertexFormatAttribIndeces[flagIndex] = glGetAttribLocation(program, attributeName);
+        if(flagIndex != -1)
+        {
+            int32 attributeLocationIndex = glGetAttribLocation(program, attributeName);
+            vertexFormatAttribIndeces[flagIndex] = attributeLocationIndex;
+            requiredVertexFormat |= 1<<flagIndex;
+        }
     }
     
     RENDER_VERIFY(glGetProgramiv(program, GL_ACTIVE_UNIFORMS, &activeUniforms));
@@ -409,7 +425,7 @@ void Shader::RecompileInternal(BaseObject * caller, void * param, void *callerDa
         
         if(IsAutobindUniform(shaderSemantic))
         {
-            //Logger::FrameworkDebug(Format("%s %d", attrName.c_str(), shaderSemantic).c_str());
+            Logger::FrameworkDebug(Format("Autobind: %s %d", attrName.c_str(), shaderSemantic).c_str());
             autobindUniformCount++;
         }
         
@@ -971,7 +987,7 @@ GLint Shader::CompileShader(GLuint *shader, GLenum type, GLint count, const GLch
         RENDER_VERIFY(glGetShaderInfoLog(*shader, 4096, &logLength, log));
         if (logLength)
         {
-            Logger::FrameworkDebug("Shader compile log:\n%s", log);
+            Logger::Error("Shader compile log:\n%s", log);
         }
     }
 #endif
@@ -1003,6 +1019,7 @@ void Shader::Bind()
 {
     if (activeProgram != program)
     {
+        //Logger::FrameworkDebug(Format("Bind: %d", program).c_str());
         RENDERER_UPDATE_STATS(shaderBindCount++);
         RENDER_VERIFY(glUseProgram(program));
         activeProgram = program;
@@ -1051,19 +1068,6 @@ void Shader::BindDynamicParameters()
                 }
                 break;
             }
-            case PARAM_WORLD_VIEW_TRANSLATE:
-            {
-                RenderManager::Instance()->ComputeWorldViewMatrixIfRequired();
-                pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(PARAM_WORLD_VIEW);
-                if (_updateSemantic != currentUniform->updateSemantic)
-                {
-                    RENDERER_UPDATE_STATS(dynamicParamUniformBindCount++);
-                    Matrix4 * world = (Matrix4*)RenderManager::GetDynamicParam(PARAM_WORLD_VIEW);
-                    SetUniformValueByUniform(currentUniform, world->GetTranslationVector());
-                    currentUniform->updateSemantic = _updateSemantic;
-                }
-                break;
-            }
             case PARAM_WORLD_SCALE:
             {
                 pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(PARAM_WORLD);
@@ -1085,6 +1089,19 @@ void Shader::BindDynamicParameters()
                 {
                     RENDERER_UPDATE_STATS(dynamicParamUniformBindCount++);
                     Matrix4 * proj = (Matrix4*)RenderManager::GetDynamicParam(PARAM_PROJ);
+                    SetUniformValueByUniform(currentUniform, *proj);
+                    currentUniform->updateSemantic = _updateSemantic;
+                }
+                break;
+            }
+            case PARAM_INV_WORLD_VIEW:
+            {
+                RenderManager::Instance()->ComputeInvWorldViewMatrixIfRequired();
+                pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(PARAM_INV_WORLD_VIEW);
+                if (_updateSemantic != currentUniform->updateSemantic)
+                {
+                    RENDERER_UPDATE_STATS(dynamicParamUniformBindCount++);
+                    Matrix4 * proj = (Matrix4*)RenderManager::GetDynamicParam(PARAM_INV_WORLD_VIEW);
                     SetUniformValueByUniform(currentUniform, *proj);
                     currentUniform->updateSemantic = _updateSemantic;
                 }
@@ -1172,27 +1189,98 @@ void Shader::BindDynamicParameters()
             case PARAM_CAMERA_POS:
             case PARAM_CAMERA_DIR:
             case PARAM_CAMERA_UP:
+            case PARAM_LIGHT0_COLOR:
+            case PARAM_LIGHT0_AMBIENT_COLOR:
             {
                 pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(currentUniform->shaderSemantic);
                 if (_updateSemantic != currentUniform->updateSemantic)
                 {
                     RENDERER_UPDATE_STATS(dynamicParamUniformBindCount++);
-                    
-                    Vector3 * camParam = (Vector3*)RenderManager::GetDynamicParam(currentUniform->shaderSemantic);
-                    //RENDER_VERIFY(glUniform3fv(currentUniform->location, 1, (float*)camParam));
-                    SetUniformValueByUniform(currentUniform, *camParam);
-
+                    Vector3 * param = (Vector3*)RenderManager::GetDynamicParam(currentUniform->shaderSemantic);
+                    SetUniformValueByUniform(currentUniform, *param);
                     currentUniform->updateSemantic = _updateSemantic;
                 }
                 break;
             }
-
+            case PARAM_LIGHT0_POSITION:
+            {
+                pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(currentUniform->shaderSemantic);
+                if (_updateSemantic != currentUniform->updateSemantic)
+                {
+                    RENDERER_UPDATE_STATS(dynamicParamUniformBindCount++);
+                    Vector4 * param = (Vector4*)RenderManager::GetDynamicParam(currentUniform->shaderSemantic);
+                    SetUniformValueByUniform(currentUniform, *param);
+                    currentUniform->updateSemantic = _updateSemantic;
+                }
+                break;
+            }
+            case PARAM_WORLD_VIEW_OBJECT_CENTER:
+            {
+                RenderManager::Instance()->ComputeWorldViewMatrixIfRequired();
+                pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(PARAM_WORLD_VIEW);
+                if (_updateSemantic != currentUniform->updateSemantic)
+                {
+                    AABBox3 * objectBox = (AABBox3*)RenderManager::GetDynamicParam(PARAM_LOCAL_BOUNDING_BOX);
+                    Matrix4 * worldView = (Matrix4 *)RenderManager::GetDynamicParam(PARAM_WORLD_VIEW);
+                    Vector3 param = objectBox->GetCenter() * (*worldView);
+                    SetUniformValueByUniform(currentUniform, param);
+                    currentUniform->updateSemantic = _updateSemantic;
+                }
+                break;
+            }
+            case PARAM_BOUNDING_BOX_SIZE:
+            {
+                pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(PARAM_LOCAL_BOUNDING_BOX);
+                if (_updateSemantic != currentUniform->updateSemantic)
+                {
+                    AABBox3 * objectBox = (AABBox3*)RenderManager::GetDynamicParam(PARAM_LOCAL_BOUNDING_BOX);
+                    Vector3 param = objectBox->GetSize();
+                    SetUniformValueByUniform(currentUniform, param);
+                    currentUniform->updateSemantic = _updateSemantic;
+                }
+                break;
+            }
+            case PARAM_SPEED_TREE_LEAFS_OSCILLATION:
+            case PARAM_SPEED_TREE_TRUNK_OSCILLATION:
+            {
+                pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(currentUniform->shaderSemantic);
+                if (_updateSemantic != currentUniform->updateSemantic)
+                {
+                    Vector2 * param = (Vector2*)RenderManager::GetDynamicParam(currentUniform->shaderSemantic);
+                    SetUniformValueByUniform(currentUniform, *param);
+                    currentUniform->updateSemantic = _updateSemantic;
+                }
+                break;
+            }
+            case PARAM_SPEED_TREE_LIGHT_SMOOTHING:
+            {
+                pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(currentUniform->shaderSemantic);
+                if (_updateSemantic != currentUniform->updateSemantic)
+                {
+                    float32 * param = (float32*)RenderManager::GetDynamicParam(currentUniform->shaderSemantic);
+                    SetUniformValueByUniform(currentUniform, *param);
+                    currentUniform->updateSemantic = _updateSemantic;
+                }
+                break;
+            }
+            case PARAM_SPHERICAL_HARMONICS:
+            {
+                pointer_size _updateSemantic = GET_DYNAMIC_PARAM_UPDATE_SEMANTIC(PARAM_SPHERICAL_HARMONICS);
+                if (_updateSemantic != currentUniform->updateSemantic)
+                {
+                    Vector3 * param = (Vector3*)RenderManager::GetDynamicParam(PARAM_SPHERICAL_HARMONICS);
+                    SetUniformValueByUniform(currentUniform, Shader::UT_FLOAT_VEC3, currentUniform->size, param);
+                    currentUniform->updateSemantic = _updateSemantic;
+                }
+                break;
+            }
             case PARAM_COLOR:
             {
                 const Color & c = RenderManager::Instance()->GetColor();
                 SetUniformColor4ByUniform(currentUniform, c);
                 break;
             }
+            
             case PARAM_GLOBAL_TIME:
             {
                 if (currentUniform->updateSemantic != Core::Instance()->GetGlobalFrameIndex())
@@ -1308,6 +1396,8 @@ void Shader::Lost()
     program = 0;
     activeAttributes = 0;
     activeUniforms = 0;
+    
+    ReleaseShaderData(false);
 }
 
 void Shader::Invalidate()
